Lab 3 report, inverse function and matrix mult.
An inverse function is in inverse.c which we could use if we still need it. 

The report for part a and b is done; however, please check it over and make sure it is correct (especially the alg steps for each desirable property) and make sure b is correct, because I believe the logic in is sound, but I'm not sure if I am missing anything.

I'm not sure if we still need a matrix multiplication function (here is one that could be modified if we still do.